+++ CHANGELOG.md
@@ -612,3 +612,4 @@
 * Generalize `Hoistable` typeclass
 * Define experimental notation for map methods of `Covariant` typeclass
 * Define experimental notation for map methods of `Contravariant` typeclass
+* Define `Tape` type synonymous for List and Stream zippers